decoupling, refactoring and a fancy treat in the form of weighted fuzzy scoring 🍬
