+++ api/process.py
@@ -1,17 +1,14 @@
 from flask import request
-from .utils import (
+from .utils.fuzzy import fuzzyfy
+from .utils.diff import diff_frames
+from .mod import contacts, corporations
+from .utils.common import (
     export,
-    fuzzyfy,
     read_csv,
     filename,
     parse_list,
-    diff_frames,
     condo_coop_mask,
-    prepare_contacts,
-    post_process_contacts,
-    prepare_corporation_count,
 )
-
 
 RPC = dict()
 
@@ -27,7 +24,7 @@ def corporation_count():
     buildings_file = request.files.get('buildings')
     building_cols = parse_list(request.form.get('building-columns'))
     filter_keywords = parse_list(request.form.get('filter-keywords'))
-    df = prepare_corporation_count(
+    df = corporations.prepare(
         contacts_file,
         buildings_file,
         building_cols,
@@ -60,16 +57,16 @@ def compare_contacts():
         'LastName',
     ]
 
-    contacts_old = prepare_contacts(contacts_old, index)
-    contacts_new = prepare_contacts(contacts_new, index, new=True)
+    contacts_old = contacts.prepare(contacts_old, index)
+    contacts_new = contacts.prepare(contacts_new, index, new=True)
 
     old_rids = contacts_old['RegistrationID'].copy()
 
     dfc = diff_frames(
         contacts_old,
         contacts_new,
-        ignore_cols=index,
-        show_atleast=[*index, 'BusinessZip'],
+        ignore_cols=(*index,),
+        show_atleast=(*index, 'BusinessZip'),
         removed_mask=condo_coop_mask)
 
     del contacts_old, contacts_new
@@ -82,10 +79,10 @@ def compare_contacts():
                              dtype={'BuildingID': 'UInt32',
                                     'RegistrationID': 'UInt32'})
 
-        dfc = post_process_contacts(
+        dfc = contacts.post_process(
             dfc, buildings, old_rids=old_rids,
-            col_order={'first': ['ChangeType', *index],
-                       'last': ['BusinessZip', 'Zip', 'ZipMatch']})
+            col_order={'first': ('ChangeType', *index),
+                       'last': ('BusinessZip', 'Zip', 'ZipMatch')})
 
         del buildings, old_rids
