+++ main.py
@@ -1,9 +1,9 @@
-from flask import Flask, request, abort, render_template
-from gevent.pywsgi import WSGIServer
-from flask_compress import Compress
-from api.process import RPC
 import traceback
 from gevent import monkey
+from gevent.pywsgi import WSGIServer
+from flask_compress import Compress
+from flask import Flask, request, abort, render_template
+from api.process import RPC
 monkey.patch_all()
