+++ templates/main.html
@@ -57,6 +57,7 @@
         min="0"
         max="100"
         value="90"
+        step="any"
         type="number"
         name="similarity"
         title="Value between 0-100. For e.g. 90 will group together names that are 90% similar"
